Rename suiteexecution package in api (#50)
