+++ client/client-core/src/main/java/com/cognifide/aet/common/FatalErrorStatusProcessor.java
@@ -18,7 +18,7 @@
 package com.cognifide.aet.common;
 
 import com.cognifide.aet.communication.api.exceptions.AETException;
-import com.cognifide.aet.communication.api.suiteexecution.SuiteStatusResult;
+import com.cognifide.aet.communication.api.execution.SuiteStatusResult;
 
 class FatalErrorStatusProcessor implements StatusProcessor {
