+++ client/client-core/src/main/java/com/cognifide/aet/common/ProcessingErrorStatusProcessor.java
@@ -18,7 +18,7 @@
 package com.cognifide.aet.common;
 
 import com.cognifide.aet.communication.api.exceptions.AETException;
-import com.cognifide.aet.communication.api.suiteexecution.SuiteStatusResult;
+import com.cognifide.aet.communication.api.execution.SuiteStatusResult;
 import com.jcabi.log.Logger;
 
 public class ProcessingErrorStatusProcessor implements StatusProcessor {
