+++ client/client-core/src/test/java/com/cognifide/aet/common/ProcessingErrorMessageProcessorTest.java
@@ -18,7 +18,7 @@
 package com.cognifide.aet.common;
 
 import com.cognifide.aet.communication.api.exceptions.AETException;
-import com.cognifide.aet.communication.api.suiteexecution.SuiteStatusResult;
+import com.cognifide.aet.communication.api.execution.SuiteStatusResult;
 
 import org.junit.Before;
 import org.junit.Test;
