+++ test-executor/src/main/java/com/cognifide/aet/executor/SuiteStatusServlet.java
@@ -19,7 +19,7 @@ package com.cognifide.aet.executor;
 
 import com.google.gson.Gson;
 
-import com.cognifide.aet.communication.api.suiteexecution.SuiteStatusResult;
+import com.cognifide.aet.communication.api.execution.SuiteStatusResult;
 
 import org.apache.commons.lang.StringUtils;
 import org.apache.felix.scr.annotations.Activate;
